Add files via upload
In this project, the program asks the user to enter the number of subjects and then prompts for the grade of each subject. It calculates the total grade by summing up all the grades and then calculates the average grade by dividing the total by the number of subjects. Finally, it displays the average grade to the user.